requirements.txt
altgraph==0.17.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
idna==3.3
lockers-cl==0.0.1
pycairo==1.21.0
PyGObject==3.42.2
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.2
pyserial==3.5
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11

